Move threadpool file
