+++ libvcx/src/api/issuer_credential.rs
@@ -7,7 +7,7 @@ use crate::{connection, credential_def, issuer_credential, settings};
 use crate::error::prelude::*;
 use crate::abi_utils::cstring::CStringUtils;
 use crate::utils::error;
-use crate::utils::runtime::execute;
+use crate::abi_utils::runtime::execute;
 
 /*
     The API represents an Issuer side in credential issuance process.
