+++ libvcx/src/api/utils.rs
@@ -14,7 +14,7 @@ use crate::libindy::utils::payments;
 use crate::utils::constants::*;
 use crate::abi_utils::cstring::CStringUtils;
 use crate::utils::error;
-use crate::utils::runtime::execute;
+use crate::abi_utils::runtime::execute;
 use crate::utils::provision::AgentProvisionConfig;
 
 #[derive(Deserialize, Debug, Clone)]
