+++ libvcx/src/init.rs
@@ -1,13 +1,13 @@
 use indy;
 use indy::ErrorCode;
-use indy_sys::WalletHandle;
 use indy::future::Future;
+use indy_sys::WalletHandle;
 
-use crate::{settings, utils};
-use crate::error::{VcxErrorExt, VcxError, VcxErrorKind, VcxResult};
+use crate::{abi_utils, settings, utils};
+use crate::abi_utils::runtime::ThreadpoolConfig;
+use crate::error::{VcxError, VcxErrorExt, VcxErrorKind, VcxResult};
 use crate::libindy::utils::pool::{create_pool_ledger_config, open_pool_ledger};
-use crate::libindy::utils::wallet::{build_wallet_config, build_wallet_credentials, set_wallet_handle, WalletConfig, IssuerConfig};
-use crate::utils::runtime::ThreadpoolConfig;
+use crate::libindy::utils::wallet::{build_wallet_config, build_wallet_credentials, IssuerConfig, set_wallet_handle, WalletConfig};
 use crate::utils::provision::AgencyClientConfig;
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
@@ -20,7 +20,7 @@ pub struct PoolConfig {
 pub fn init_threadpool(config: &str) -> VcxResult<()> {
     let config: ThreadpoolConfig = serde_json::from_str(config)
         .map_err(|err| VcxError::from_msg(VcxErrorKind::InvalidJson, format!("Failed to deserialize threadpool config {:?}, err: {:?}", config, err)))?;
-    utils::runtime::init_runtime(config);
+    abi_utils::runtime::init_runtime(config);
     Ok(())
 }
