+++ libvcx/src/utils/devsetup.rs
@@ -2,20 +2,21 @@ use std::fs;
 use std::sync::Once;
 
 use crate::{libindy, settings, utils};
+use crate::abi_utils::runtime;
+use crate::abi_utils::runtime::ThreadpoolConfig;
 use crate::agency_client::mocking::AgencyMockDecrypted;
+use crate::init::PoolConfig;
 use crate::libindy::utils::pool::reset_pool_handle;
 use crate::libindy::utils::pool::tests::{create_test_ledger_config, delete_test_pool, open_test_pool};
-use crate::libindy::utils::wallet::{close_main_wallet, create_indy_wallet, delete_wallet, reset_wallet_handle, WalletConfig, create_and_open_as_main_wallet};
+use crate::libindy::utils::wallet::{close_main_wallet, create_and_open_as_main_wallet, create_indy_wallet, delete_wallet, reset_wallet_handle, WalletConfig};
 use crate::settings::set_testing_defaults;
-use crate::utils::{get_temp_dir_path, runtime};
 use crate::utils::constants;
+use crate::utils::devsetup_agent::test::{Alice, Faber, TestAgent};
 use crate::utils::file::write_file;
+use crate::utils::get_temp_dir_path;
 use crate::utils::logger::LibvcxDefaultLogger;
 use crate::utils::object_cache::ObjectCache;
 use crate::utils::plugins::init_plugin;
-use crate::utils::runtime::ThreadpoolConfig;
-use crate::init::PoolConfig;
-use crate::utils::devsetup_agent::test::{Faber, Alice, TestAgent};
 
 pub struct SetupEmpty; // clears settings, setups up logging
