+++ libvcx/src/utils/mod.rs
@@ -38,7 +38,6 @@ pub mod error;
 pub mod constants;
 pub mod openssl;
 pub mod json;
-pub mod runtime;
 pub mod uuid;
 pub mod author_agreement;
 pub mod qualifier;
